Added : Encrypt & Decrypt data password rtdb; Fix Bug : koordinator hapus wartawan malah ke login wartawan, Hapus Wartawan (auth, rtdb, storage) berhasil but ada  kendala disaat load hapus foto profil -> force close

+++ app/src/main/java/com/example/sikobeh/WartawanAdapter.java
@@ -2,6 +2,7 @@ package com.example.sikobeh;
 
 import android.content.Context;
 import android.content.Intent;
+import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,15 +14,19 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.bumptech.glide.Glide;
+import com.bumptech.glide.load.DataSource;
+import com.bumptech.glide.load.engine.GlideException;
+import com.bumptech.glide.request.RequestListener;
+import com.bumptech.glide.request.target.Target;
 import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -30,11 +35,8 @@ import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
-import com.squareup.picasso.Callback;
-import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
-import java.util.concurrent.TimeUnit;
 
 import cn.pedant.SweetAlert.SweetAlertDialog;
 import de.hdodenhof.circleimageview.CircleImageView;
@@ -46,6 +48,8 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
     ArrayList<User> list;
     FirebaseAuth fAuth;
     StorageReference storageReference;
+    String AES = "AES";
+    String decryptedPass;
 
     public static Boolean clickBerita = false;
 
@@ -70,11 +74,17 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
         holder.pnumber.setText(user.getPnumber());
         String value = user.getUid();
         String email = user.getEmail();
-        String pass = user.getPassword();
+        try {
+            decryptedPass = RegisterUser.decrypt(user.getPassword(), "rasul19");
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        String pass = decryptedPass;
 
         fAuth = FirebaseAuth.getInstance();
         storageReference = FirebaseStorage.getInstance().getReference();
 
+        /*
         StorageReference profileRef = storageReference.child("users/"+value+"/profile.jpg");
         profileRef.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
             @Override
@@ -100,6 +110,25 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
                 holder.bimage.setVisibility(View.VISIBLE);
             }
         });
+         */
+
+        Glide.with(holder.bimage.getContext()).load(user.getImageurl())
+                .listener(new RequestListener<Drawable>() {
+                    @Override
+                    public boolean onLoadFailed(@Nullable GlideException e, Object model, Target<Drawable> target, boolean isFirstResource) {
+                        holder.bimage.setVisibility(View.VISIBLE);
+                        holder.progressBar.setVisibility(View.GONE);
+                        return false;
+                    }
+
+                    @Override
+                    public boolean onResourceReady(Drawable resource, Object model, Target<Drawable> target, DataSource dataSource, boolean isFirstResource) {
+                        holder.bimage.setVisibility(View.VISIBLE);
+                        holder.progressBar.setVisibility(View.GONE);
+                        return false;
+                    }
+                })
+                .error(R.drawable.ic_baseline_account_circle_24_white).into(holder.bimage);
 
         String value2 = holder.fullname.getText().toString().trim();
 
@@ -174,12 +203,7 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
                         profileRef.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
                             @Override
                             public void onSuccess(Uri uri) {
-                                profileRef.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                                    @Override
-                                    public void onSuccess(Void unused) {
-
-                                    }
-                                });
+                                profileRef.delete();
                             }
                         });
                         // -> DataBerita
@@ -192,9 +216,13 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
                                     dataSnapshot.getRef().removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
                                         @Override
                                         public void onComplete(@NonNull Task<Void> task) {
-                                            String urlGambar = dataSnapshot.child("beritaurl").getValue().toString();
-                                            StorageReference ref2 = FirebaseStorage.getInstance().getReferenceFromUrl(urlGambar);
-                                            ref2.delete();
+                                            if (task.isSuccessful()){
+                                                String urlGambar = dataSnapshot.child("beritaurl").getValue().toString();
+                                                StorageReference ref2 = FirebaseStorage.getInstance().getReferenceFromUrl(urlGambar);
+                                                ref2.delete();
+                                            } else {
+                                                Toasty.error(context, task.getException().getMessage(), Toast.LENGTH_LONG, true).show();
+                                            }
                                         }
                                     });
                                 }
@@ -205,46 +233,39 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
 
                             }
                         });
-                        final FirebaseUser user;
                         // Delete Authentication
                         FirebaseAuth.getInstance().signOut(); // Hapus jika poin 3 kendala terpenuhi
-                        FirebaseAuth.getInstance().signInWithEmailAndPassword(email.trim(),pass.trim()).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                        FirebaseAuth.getInstance().signInWithEmailAndPassword(email, pass).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
                             @Override
                             public void onComplete(@NonNull Task<AuthResult> task) {
                                 if (task.isSuccessful()){
-                                    FirebaseAuth.getInstance().getCurrentUser().delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                                    FirebaseAuth.getInstance().getCurrentUser().delete().addOnSuccessListener(new OnSuccessListener<Void>() {
                                         @Override
-                                        public void onComplete(@NonNull Task<Void> task) {
-                                            if (task.isSuccessful()){
-                                                dialog.setTitle("Terhapus!");
-                                                dialog.setContentText("Wartawan Berhasil Terhapus!")
-                                                       .setConfirmText("Oke")
-                                                       .setConfirmClickListener(null)
-                                                       .changeAlertType(SweetAlertDialog.SUCCESS_TYPE);
-                                            } else {
-                                                Toasty.error(context, task.getException().getMessage(), Toast.LENGTH_LONG, true).show();
-                                            }
+                                        public void onSuccess(Void aVoid) {
+                                            dialog.setTitle("Terhapus!");
+                                            dialog.setContentText("Wartawan Berhasil Terhapus!")
+                                                    .setConfirmText("Oke")
+                                                    .setConfirmClickListener(null)
+                                                    .changeAlertType(SweetAlertDialog.SUCCESS_TYPE);
+
+                                            // delete Realtime Database
+                                            DatabaseReference db = FirebaseDatabase.getInstance().getReference("Users").child(value);
+                                            db.getRef().removeValue().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                @Override
+                                                public void onSuccess(Void aVoid) {
+                                                    FirebaseAuth.getInstance().signOut(); // Hapus jika poin 3 kendala terpenuhi
+                                                    Toasty.info(context, "Memuat kembali data..", Toast.LENGTH_SHORT, true).show();
+                                                    //notifyDataSetChanged();
+                                                }
+                                            });
                                         }
                                     });
-                                    FirebaseAuth.getInstance().signOut(); // Hapus jika poin 3 kendala terpenuhi
                                 }
                                 else {
                                     Toasty.error(context, task.getException().getMessage(), Toast.LENGTH_LONG, true).show();
                                 }
                             }
                         });
-
-                        // delete Realtime Database
-                        DatabaseReference db = FirebaseDatabase.getInstance().getReference("Users").child(value);
-                        db.removeValue();
-                        try {
-                            TimeUnit.MILLISECONDS.sleep(500);
-                        } catch (InterruptedException e) {
-                            e.printStackTrace();
-                        }
-                        Toasty.info(context, "Memuat kembali data..", Toast.LENGTH_SHORT, true).show();
-                        list.clear();
-                        notifyDataSetChanged();
                     }
                 }).setCancelClickListener(new SweetAlertDialog.OnSweetClickListener() {
             @Override
@@ -253,6 +274,5 @@ public class WartawanAdapter extends RecyclerView.Adapter<WartawanAdapter.MyView
             }
         }).show();
 
-        }
-
+    }
 }
